pyproject.toml
[project]
name = "mimicking-mindsets"
version = "0.1.0"
description = "Multi-agent AI system simulating Turkish intellectuals Erol Güngör and Cemil Meriç"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core AI/ML Dependencies - CUDA enabled
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "sentence-transformers>=2.2.2",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langgraph>=0.2.0",
    "langsmith>=0.1.0",
    "qdrant-client>=1.14.3",
    "numpy>=2.1.0",
    "duckduckgo-search>=3.9.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
    # FastAPI backend server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "requests>=2.31.0",
    "datasets>=2.14.0",
    "pandas>=2.0.0",
    "tabulate>=0.9.0",
    "ragas>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "evaluation", "web-interface", "knowledge-base", "tests"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
